requirements.txt
amqp==2.6.1
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
Django==3.1.1
django-cors-headers==3.5.0
django-notifs==2.6.3
django-rest-registration==0.5.6
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
idna==2.10
isort==5.4.2
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
ndg-httpsclient==0.5.1
pika==1.1.0
Pillow==7.2.0
psycopg2-binary==2.8.5
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typing==3.7.4.3
urllib3==1.25.10
uWSGI==2.0.19.1
vine==1.3.0
wrapt==1.12.1
